+++ src/main/resources/templates/borrowlog.html
@@ -25,6 +25,7 @@
 						alt="Rakuten_sec_mark" class="main_logo"> </a></div>
 			<div class="header_title"> 楽天図書館 </div>
 			<div class="header_btn_box">
+				<!--ログイン時出力される（お知らせ遷移用アイコン、ユーザー別アイコン）-->
 				<table class="header_table">
 					<th>
 					<td> <img src="img/alarm_icon.png" class="alert_icon"> </td>
